RAI Ymca
We are a group of college students like you, who are attending online classes due to covid and struggling to find notes and guidance on online courses.
Why did we come with the idea to make this website?

It all started when we got fewer marks in JEE Mains. After attending various physical counsellings from college to college for admission, looking for the best possible branch in the best possible college, we got the newly-launched Robotics and Artificial Intelligence branch at YMCA, Faridabad. We were happy because, in those competitive days, it was difficult for a student with a low JEE Main rank to get into a good college.
But who knew that in our journey of B.Tech in the newly-launched branch, we would not have any senior or fully defined curriculum? We were stuck at home and were attending online classes due to Covid. With no physical interaction with classmates, seniors or professors, we were not getting proper notes and guidance on online courses. Studying from home was frustrating us. So, we decided that we would not let our juniors suffer the same kind of pain and came up with the idea of ​​creating a website for our newly launched branch. We thought of organizing our complete data, i.e. syllabus, notes and online resources, on that website.
--> It's a non-profit website created to help college students with study materials, guidance on online courses, and much more!